+++ ClickHouse.Core/Models/DateCount.cs
@@ -3,5 +3,5 @@ namespace ClickHouse.Core.Models;
 public record DateCount
 {
     public DateTime Date { get; set; }
-    public int Count { get; set; }
+    public ulong Count { get; set; }
 };
