+++ ClickHouse.Core/Models/SensorCount.cs
@@ -3,5 +3,5 @@ namespace ClickHouse.Core.Models;
 public record SensorCount
 {
     public string SensorType { get; set; }
-    public int Count { get; set; }
+    public ulong Count { get; set; }
 }
